// src/store/consumer/basic/building/types/building.ts
export interface BuildingTypeRef {
  building_type_id: string;
  name: string;
  createdAt: string;
  updatedAt: string;
}

export interface BuildingSubTypeRef {
  building_sub_type_id: string;
  name: string;
  building_type_id: string;
  createdAt: string;
  updatedAt: string;
}

export interface UtilityStateRef {
  id: string;
  createdAt: string;
  state_id: string;
  utility_company_id: string;
  state: {
    state_id: string;
    name: string;
  };
}

export interface UtilityCompanyRef {
  utility_company_id: string;
  utility_company_name: string;
  country_id: string;
  states: UtilityStateRef[];
}

export interface CommodityRef {
  commodity_id: string;
  name: string;
  createdAt: string;
  updatedAt: string;
}

export interface UserBuildingUtility {
  accountNumber: string;
  unit: string;
  acceptTermsAndConditions: boolean;
  acceptanceDate: string;
  utility: UtilityCompanyRef;
  commodity: CommodityRef;
  createdAt: string;
  updatedAt: string;
}

export interface BuildingCardSummary {
  rooms: number;
  evs: number;
  energy_usage: number;
  type: string;
  sub_type: string;
}

export interface UserBuilding {
  user_building_details_id: string;
  building_name: string;
  postalCode: string;
  city: string;
  building_type: BuildingTypeRef;
  building_sub_type: BuildingSubTypeRef;
  streetAddress: string;
  createdAt: string;
  updatedAt: string;
  noOfOccupants: string;
  user_building_utility: UserBuildingUtility[];

  building_type_name: string;
  building_sub_type_name: string;
  type: string;
  sub_type: string;
  rooms_count: number;
  room_count: number;
  evs_count: number;
  ev_count: number;
  energy_usage: number;
  energyUsage: number;
  card_summary: BuildingCardSummary;
}